Add a package with a file operations
